Making xAOD::TEvent a bit more silent when using it through the IProxyDict interface (xAODRootAccess-00-01-62)
	* Silencing some more warnings through the IProxyDict interface.
	  Not to flood the xAODChecker logs with irrelevant messages
	  too much.
	* Tagging as xAODRootAccess-00-01-62

2016-11-16 Attila Krasznahorkay <Attila.Krasznahorkay@cern.ch>
	* Merged the changes back into the trunk.
	* And then changed the implementation quite significantly. As the
	  previous one had all sorts of problems that were revealed by
	  the transient tree tests. In the new implementation the auxiliary
	  stores should now be set up correctly when retrieving objects
	  through the IProxyDict interface.
	* Tagging as xAODRootAccess-00-01-61

2016-11-10 Attila Krasznahorkay <Attila.Krasznahorkay@cern.ch>
	* Implemented the rest of Scott's modifications, to make the code
	  work in the offline environment. (Tested on top of devval.)
	* Tried to make the code just a little more clear on ownership
	  issues along the way.

...
(Long ChangeLog diff - truncated)


Former-commit-id: abe92cfb269a12ca948aabf5df09d2b935b36004

+++ Control/xAODRootAccess/Root/TPrimitiveAuxBranchManager.cxx
@@ -2,7 +2,7 @@
   Copyright (C) 2002-2017 CERN for the benefit of the ATLAS collaboration
 */
 
-// $Id: TPrimitiveAuxBranchManager.cxx 595278 2014-05-03 09:31:05Z krasznaa $
+// $Id: TPrimitiveAuxBranchManager.cxx 783066 2016-11-08 19:39:32Z ssnyder $
 
 // ROOT include(s):
 #include <TBranch.h>
@@ -34,7 +34,7 @@ namespace xAOD {
          m_holder = new THolder( *parent.m_holder );
       }
       if( parent.m_vector ) {
-         m_vector = SG::AuxTypeRegistry::instance().makeVector( m_auxId, 0, 0 );
+         m_vector = SG::AuxTypeRegistry::instance().makeVector( m_auxId, (size_t)0, (size_t)0 );
          m_vector->resize( 1 );
       }
    }
@@ -70,7 +70,7 @@ namespace xAOD {
       m_auxId = rhs.m_auxId;
       if( m_vector ) delete m_vector;
       if( rhs.m_vector ) {
-         m_vector = SG::AuxTypeRegistry::instance().makeVector( m_auxId, 0, 0 );
+         m_vector = SG::AuxTypeRegistry::instance().makeVector( m_auxId, (size_t)0, (size_t)0 );
          m_vector->resize( 1 );
       } else {
          m_vector = 0;
@@ -145,7 +145,7 @@ namespace xAOD {
       // Otherwise let's create a default object:
       m_isSet = kTRUE;
       if( ! m_vector ) {
-         m_vector = SG::AuxTypeRegistry::instance().makeVector( m_auxId, 0, 0 );
+         m_vector = SG::AuxTypeRegistry::instance().makeVector( m_auxId, (size_t)0, (size_t)0 );
          m_vector->resize( 1 );
       }
       // ...and use it to fill the current event:
